% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppt_ph_annotate.R
\name{add_annotated_layouts}
\alias{add_annotated_layouts}
\title{Add layout slide with annotated placeholders}
\usage{
add_annotated_layouts(x, layouts = NULL, remove_slides = FALSE, ...)
}
\arguments{
\item{x}{A \code{rpptx} object.}

\item{layouts}{Names or indexes of layouts to annotate. If \code{NULL} (default), all layouts are used.}

\item{remove_slides}{Remove already existing slides? (default \code{FALSE}). If \code{TRUE}, will only keep annotated layouts.}

\item{...}{Passed on to \code{\link[=phs_annotate]{phs_annotate()}} for finetuning placeholder's appearance.}
}
\value{
\code{rpptx} object with annotated layout slides added.
}
\description{
Generates a new slide with annotated placeholders (\code{label}, \code{id}, \verb{type + type index})
for each layout. This overview is helpful when calling \code{\link[=phs_with]{phs_with()}}, \code{\link[=ph_with]{ph_with()}}, or
the \verb{ph_location_*()} functions.
}
\seealso{
Other ppt_info:
\code{\link[=annotate_base]{annotate_base()}},
\code{\link[=phs_annotate]{phs_annotate()}},
\code{\link[=plot_layout_properties]{plot_layout_properties()}},
\code{\link[=slide_size]{slide_size()}}
}
\concept{ppt_info}
