% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppt_ph_annotate.R
\name{annotate_base}
\alias{annotate_base}
\title{Layout slides with annotated placeholders}
\usage{
annotate_base(
  path = NULL,
  output_file = "annotated_layouts.pptx",
  layouts = NULL,
  ...
)
}
\arguments{
\item{path}{Path to pptx file. If \code{NULL} (default), \code{officer}'s default presentation is used.}

\item{output_file}{Path to save annotated presentation. Use \code{NULL} to suppress file generation.}

\item{layouts}{Names or indexes of layouts to annotate. If \code{NULL} (default), all layouts are used.}

\item{...}{Passed on to \code{\link[=add_annotated_layouts]{add_annotated_layouts()}} for finetuning appearance.}
}
\value{
\code{rpptx} object with annotated layouts.
}
\description{
Generates a slide with annotated placeholders (\code{label}, \code{id}, \verb{type + type index})
for each layout of the pptx file. This placeholder overview is helpful when calling \code{\link[=phs_with]{phs_with()}},
\code{\link[=ph_with]{ph_with()}}, or the \verb{ph_location_*()} functions.
}
\examples{
# Annotate all layouts in officer's default presentation
\dontrun{
annotate_base() # creates local file `annotated_layouts.pptx`
}

# without file creation
x <- annotate_base(output_file = NULL)
\dontrun{
print(x, preview = TRUE)}

# annotate selected layouts only (if there are too many)
x <- annotate_base(layouts = c("Title Slide", "Two Content"), output_file = NULL)
x <- annotate_base(layouts = c(1,4), output_file = NULL)  # same using layout index
\dontrun{
print(x, preview = TRUE)}

# save annotated layouts directly to file
file_out <- tempfile(fileext = ".pptx")
annotate_base(output_file = file_out)
\dontrun{
file_open(file_out)}
}
\seealso{
Other ppt_info:
\code{\link[=add_annotated_layouts]{add_annotated_layouts()}},
\code{\link[=phs_annotate]{phs_annotate()}},
\code{\link[=plot_layout_properties]{plot_layout_properties()}},
\code{\link[=slide_size]{slide_size()}}
}
\concept{ppt_info}
