% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ph_location.R
\name{ph_location_fullsize}
\alias{ph_location_fullsize}
\title{Location of a full size element}
\usage{
ph_location_fullsize(newlabel = "", ...)
}
\arguments{
\item{newlabel}{a label to associate with the placeholder.}

\item{...}{unused arguments}
}
\description{
The function will return the location corresponding
to a full size display.
}
\examples{
doc <- read_pptx()
doc <- add_slide(doc, "Title and Content")
doc <- ph_with(doc, "Hello world", location = ph_location_fullsize())
print(doc, target = tempfile(fileext = ".pptx"))
}
\seealso{
Other functions for placeholder location:
\code{\link[=ph_location]{ph_location()}},
\code{\link[=ph_location_id]{ph_location_id()}},
\code{\link[=ph_location_label]{ph_location_label()}},
\code{\link[=ph_location_left]{ph_location_left()}},
\code{\link[=ph_location_right]{ph_location_right()}},
\code{\link[=ph_location_template]{ph_location_template()}},
\code{\link[=ph_location_type]{ph_location_type()}}
}
\concept{functions for placeholder location}
