% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppt_ph_annotate.R
\name{phs_annotate}
\alias{phs_annotate}
\title{Annotate a slide's phs with type, label, and id}
\usage{
phs_annotate(
  x,
  ...,
  .slide_idx = NULL,
  .font_size = NA,
  .font_color = NA,
  .keys = TRUE,
  .bg = NA,
  .line = sp_line(lwd = 1, color = "grey", lty = "dash")
)
}
\arguments{
\item{x}{A \code{rpptx} object.}

\item{...}{Subset of phs to annotate (optional). Either short form location syntax (see \code{\link[=phs_with]{phs_with()}} for details) or a
\code{\link[=ph_location]{ph_location()}} object.}

\item{.slide_idx}{Indexes of slides to process. Default (\code{NULL}) is the current slide. Use \code{"all"} to process all
slides.}

\item{.font_size}{Font size. \code{NA} (default) is the ph's default font size. Use a named vector of
length 3 for different font sizes (e.g. \code{c(label = 10, type = NA, id = 15)}). Set \code{0} to omit}

\item{.font_color}{Font color(s). \code{NA} (default) is the ph's default color. Use a named vector of
length 3 for different colors (e.g. \code{c(label = "red", type = "blue", id = "darkgreen")}).}

\item{.keys}{Add keys to info (\verb{id=, type=, label=}) (default is \code{TRUE}).}

\item{.bg}{Background color as hex value or valid R color.}

\item{.line}{Line around placeholder. Either a \code{\link[=sp_line]{sp_line()}} or a single color value (hex or valid R color).}
}
\description{
Draws a box with ph info (type + type index, label, and id) at one or more placeholders. While
\code{\link[=plot_layout_properties]{plot_layout_properties()}} generates an image of the slide layout, \code{\link[=phs_annotate]{phs_annotate()}} modifies the presentation
object.
}
\examples{
# Annotate all placeholders on current slide
x <- read_pptx()
x <- add_slide(x, "Title Slide")
x <- phs_annotate(x)
\dontrun{
print(x, preview = TRUE)}


# Annotate all placeholders on all slides
x <- read_pptx()
x <- add_slide(x, "Title and Content")
x <- add_slide(x, "Two Content")
x <- phs_annotate(x, .slide_idx = "all") # annotate all slides
\dontrun{
print(x, preview = TRUE)}


# Annotate some placeholders only
x <- read_pptx()
x <- add_slide(x, "Title Slide")
x <- phs_annotate(x, "Title 1") # annotate phs with label "Title 1" only
x <- add_slide(x, "Title and Content")
x <- add_slide(x, "Two Content")
x <- phs_annotate(x, "dt", "ftr", .slide_idx = 2:3) # only types "dt" and "ftr" on slides 2 and 3
\dontrun{
print(x, preview = TRUE)}


# Modify ph appearance (usually not needed) -------

## Show label only
x <- read_pptx()
x <- add_slide(x, "Title Slide")
x <- phs_annotate(x, .font_size = c(type = 0, id = 0), .keys = FALSE)
\dontrun{
print(x, preview = TRUE)}

## Modify colors and box
x <- read_pptx()
x <- add_slide(x, "Title Slide")
x <- phs_annotate(x, .font_color = c("red", "blue", "darkgreen"), .bg = "#ff000010", .line = NA)
\dontrun{
print(x, preview = TRUE)}
}
\seealso{
\code{\link[=plot_layout_properties]{plot_layout_properties()}}

Other ppt_info:
\code{\link[=add_annotated_layouts]{add_annotated_layouts()}},
\code{\link[=annotate_base]{annotate_base()}},
\code{\link[=plot_layout_properties]{plot_layout_properties()}},
\code{\link[=slide_size]{slide_size()}}
}
\concept{ppt_info}
